Review: types moved to own file, smaller improvements
